Getting names and investor relation pages
Added a function for getting not only the CIK but also the Name. Next problem is getting reliable earnings reports. Unfortunately they are posted with the SEC later than the call or conference. So hardly any edge. For now, added some function which tries to find the investor relation page but that's proving quite difficult and doesn't work yet