Different way to get dates
Added a function get_dates_v2. It does the same as the original (supposedly) but here it just uses the dates of the SEC filings. Benefits: much more historical data. Downsides: the actual release is most often a day earlier (yfinance is not accurate either any typically tells an even earlier date), there is no future (expected) dates (can be filled with yfinance).